Many fixes due to the breaking changes of the earlier commit. Cleanup of the builder symbols interface. New utility functions and documentation.

+++ benchmark/src/serialize-big.cpp
@@ -22,7 +22,7 @@ int test_vs(std::string_view xmlInput){
 
         auto tree = bld.close();
         if(!tree.has_value()){
-            std::cerr << "Error while closing the document " << tree.error() << "\n";
+            std::cerr << "Error while closing the document " << (int)tree.error() << "\n";
             return 3;
         }
